+++ services/fighterService.js
@@ -1,29 +1,32 @@
 import { fighterRepository } from '../repositories/fighterRepository.js';
+import { BaseService } from './baseService.js';
 
-class FighterService {
+class FighterService extends BaseService {
   constructor({ fighterRepository }) {
-    this.fighterRepository = fighterRepository;
+    super({
+      repository: fighterRepository,
+    });
   }
 
-  async findAll() {
-    return this.fighterRepository.getAll();
-  }
+  // async findAll() {
+  //   return this.fighterRepository.getAll();
+  // }
 
-  async findById(id) {
-    return this.fighterRepository.getOne({ id });
-  }
+  // async findById(id) {
+  //   return this.fighterRepository.getOne({ id });
+  // }
 
-  async create(data) {
-    return this.fighterRepository.create(data);
-  }
+  // async create(data) {
+  //   return this.fighterRepository.create(data);
+  // }
 
-  async update({ id, dataToUpdate }) {
-    return this.fighterRepository.update(id, dataToUpdate);
-  }
+  // async update({ id, dataToUpdate }) {
+  //   return this.fighterRepository.update(id, dataToUpdate);
+  // }
 
-  async delete(id) {
-    return this.fighterRepository.delete(id);
-  }
+  // async delete(id) {
+  //   return this.fighterRepository.delete(id);
+  // }
 }
 
 const fighterService = new FighterService({ fighterRepository });
